requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.0
fpdf>=1.7.2
pyarrow>=14.0.0
google-generativeai>=0.8.1


